worker_node/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-worker-node"
version = "0.1.0"
description = "Standalone NexusAI worker node runtime"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pynvml>=11.5.0",
    "psutil>=5.9.0",
]

[project.scripts]
nexus-worker = "nexus_worker.__main__:main"
nexus-worker-bootstrap = "nexus_worker.bootstrap:main"

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "anyio[trio]>=4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
include = ["nexus_worker*"]
